updates to structure, transport engine, and LHS
- change transport.py to model.py since this controls the Riverine model
- create a transport engine to handle advection and diffusion
- update the LHS matrix to use the registry
